Adding folders to get stuff sorted
Just learned how to actually create a folder in VS code.
